refactor(#3479): extract recovery_engine output-path/phase-policy/jsonl-extract submodules
Behavior-preserving rank-8 split of `src/services/discord/recovery_engine.rs`.
Three pure clusters move into sub-1000-prod-LoC leaf modules under
`recovery_engine/`; the tmux/lsof subprocess drivers, async drain helpers, and
provider/tui readiness probes stay in the root module because they couple to
Command/binary_resolver/tokio/cross-service calls.

- output_path_detect.rs (177): lsof-output parsing, fallback-path matching, and
  candidate-identity logic (DetectedRebindOutputPath, LsofOutputCandidate +
  impl, StaleOutputCandidate, candidate_identity, normalize_lsof_path,
  candidate_matches_fallback, parse_lsof_output_candidates,
  detect_rebind_output_path_from_candidates).
- phase_policy.rs (120): pure recovery-phase decision predicates and the
  RecoveryPhase mapping helpers (canonical_recovery_phase, can_* gates,
  recovery_phase_after_* / recovery_phase_for_existing_inflight_rebind,
  recovery_has_post_work_ready_evidence, recovery_ready_without_output_*,
  recovery_terminal_delivery_already_committed).
- jsonl_extract.rs (137): file-offset JSONL readers
  (success_result_end_offset_after_offset, extract_response_from_output,
  extract_response_from_output_pub).

Pure move, zero logic change: moved items are re-imported via `use
self::<mod>::{...}` so all call sites stay byte-identical, and
`extract_response_from_output_pub` is re-exported so the
`recovery_engine::extract_response_from_output_pub` path stays valid for the
turn_bridge / tmux_restart_handoff external callers. recovery_engine.rs prod LoC
4090 -> 3718; baseline ratchet and change-surfaces.md freeze synced; new modules
registered via inventory regen.
