+++ scripts/audit_maintainability_giant_baseline.toml
@@ -184,7 +184,10 @@
 # caller `reregister_active_turn_from_inflight`).
 # #3034 batch-8: lowered 4094 -> 4090 (dead-code removal: dropped `assistant_turns`
 # construction sites).
-"src/services/discord/recovery_engine.rs" = 4090
+# #3479 r8: lowered 4090 -> 3718 (behavior-preserving extraction of the pure
+# output-path-detect, phase-policy, and jsonl-extract clusters into leaf modules
+# under recovery_engine/; zero logic change, call sites re-imported byte-identical).
+"src/services/discord/recovery_engine.rs" = 3718
 # #3034 batch-8: lowered 3771 -> 3770 (dead-code removal).
 # #3038 S1: +1 (3770 -> 3771) for the mechanical `shared.queued_placeholders` ->
 # `shared.queued.queued_placeholders` re-wire (one extra method-chain line) forced
